Add files via upload
A CSV file containing information about the 1,000 stars nearest to the Sun. 
Each row has seven entries. The first three entries are Cartesian coordinates (x,y,z) in a coordinate system where the sun is at the origin and the galactic plane is z=0. Subsequent entries are the star's name, radius (in units of the radius of the sun), effective temperature (Kelvin), and spectral class.